+++ Moco/Moco/MocoGoal/MocoOrientationTrackingGoal.h
@@ -35,11 +35,11 @@ using SimTK::Rotation;
 /// and integrated over the phase. This can be used to track orientation
 /// quantities in the model that don't correspond to model degrees of freedom.
 /// The reference can be provided as a trajectory of SimTK::Rotation%s
-/// representing the orientation reference data, or as a states trajectory from 
-/// which the tracked rotation reference is computed. Both rotation and states 
-/// references can be provided as a file name to a STO or CSV file (or other 
-/// file types for which there is a FileAdapter), or programmatically as a 
-/// TimeSeriesTable_<SimTK::Rotation> (for the rotation reference) or as a 
+/// representing the orientation reference data, or as a states trajectory from
+/// which the tracked rotation reference is computed. Both rotation and states
+/// references can be provided as a file name to a STO or CSV file (or other
+/// file types for which there is a FileAdapter), or programmatically as a
+/// TimeSeriesTable_<SimTK::Rotation> (for the rotation reference) or as a
 /// scalar TimeSeriesTable (for the states reference).
 ///
 /// This cost requires realization to SimTK::Stage::Position. The cost is