+++ dependencies/opensim-core.cmake
@@ -4,7 +4,7 @@
 # will invalidate its cached opensim-core installation if we change the commit.
 # This commented commit hash is not actually used in the superbuild.
 # opensim-core commit:
-# 7420c7e67aa2438b1157784f68b7fbda1b6b2498
+# d52de6c029fb6ccefa459edc23de4b04c7acb9c2
 
 AddDependency(NAME       opensim-core
               URL        ${CMAKE_SOURCE_DIR}/../opensim-core
@@ -14,17 +14,22 @@ AddDependency(NAME       opensim-core
                     -DBUILD_JAVA_WRAPPING:BOOL=${OPENSIM_JAVA_WRAPPING}
                     -DBUILD_PYTHON_WRAPPING:BOOL=${OPENSIM_PYTHON_WRAPPING}
                     -DOPENSIM_PYTHON_VERSION:STRING=${OPENSIM_PYTHON_VERSION}
-                    -DSIMBODY_HOME:PATH=${CMAKE_INSTALL_PREFIX}/simbody
-                    -DCMAKE_PREFIX_PATH:PATH=${CMAKE_INSTALL_PREFIX}/docopt
+                    -DOPENSIM_DEPENDENCIES_DIR:PATH=${CMAKE_INSTALL_PREFIX}
+                    -DWITH_BTK:BOOL=ON
                     -DOPENSIM_INSTALL_UNIX_FHS:BOOL=OFF)
 
 
 if(SUPERBUILD_opensim-core)
 
     # OpenSim's dependencies.
+    AddDependency(NAME       BTK
+                  GIT_URL    https://github.com/opensim-org/BTKCore.git
+                  GIT_TAG    6d787d0be223851a8f454f2ee8c7d9e47b84cbbe
+                  CMAKE_ARGS -DBUILD_SHARED_LIBS:BOOL=ON)
+
     AddDependency(NAME simbody
                   GIT_URL    https://github.com/simbody/simbody.git
-                  GIT_TAG    95dc0675914829f8db94cc297543d5fa340fc8df
+                  GIT_TAG    Simbody-3.7
                   CMAKE_ARGS -DBUILD_EXAMPLES:BOOL=OFF 
                              -DBUILD_TESTING:BOOL=OFF)
 
@@ -33,7 +38,7 @@ if(SUPERBUILD_opensim-core)
                   GIT_TAG    3dd23e3280f213bacefdf5fcb04857bf52e90917
                   CMAKE_ARGS -DCMAKE_DEBUG_POSTFIX:STRING=_d)
 
-    add_dependencies(opensim-core simbody docopt)
+    add_dependencies(opensim-core BTK simbody docopt)
 endif()